convert to int

+++ clone.py
@@ -22,8 +22,8 @@ train_samples, validation_samples = train_test_split(lines, test_size=0.2)
 def generator(samples, batch_size=32*3):
 	num_samples = len(samples)
 	while 1: # Loop forever so the generator never terminates
-		for offset in range(0, num_samples, batch_size/3):
-			batch_samples = samples[offset:offset+batch_size/3]
+		for offset in range(0, num_samples, int(batch_size/3)):
+			batch_samples = samples[offset:offset+int(batch_size/3)]
 
 			images = []
 			measurements = []
